Split permissions to Java-style files
