+++ src/main/scala/org/scalajs/dom/PermissionStatus.scala
@@ -1,0 +1,8 @@
+package org.scalajs.dom
+
+import scala.scalajs.js
+
+trait PermissionStatus extends EventTarget {
+  val state: PermissionState
+  var onchange: js.Function1[Event, _]
+}
