+++ src/main/scala/org/scalajs/dom/Permissions.scala
@@ -1,0 +1,7 @@
+package org.scalajs.dom
+
+import scala.scalajs.js
+
+trait Permissions extends js.Object {
+  def query(permissionDescriptor: PermissionDescriptor): js.Promise[PermissionStatus]
+}
